+++ source/html/mod.ts
@@ -1,0 +1,5 @@
+export const html = (str: TemplateStringsArray, ...args: any[]) => ({
+  isTemplate: true,
+  str,
+  args,
+});