+++ source/styled/types.ts
@@ -1,0 +1,5 @@
+export type StyledComponent<T> = {
+  (props?: T): Template & {
+    (strings: TemplateStringsArray, ...values: ElementChild[]): Template;
+  };
+};